Refactor scraper structure to make it independent of Wombat
