--- lib/generators/crawler/templates/crawler.rb.erb
@@ -1,43 +1,0 @@
-class Scrapers::<%= @agency.camelize %>::<%= @agency.camelize %>Crawler < Scrapers::ContractCrawler
-
-  protected
-  # Extract all the contract data from a contract page.
-  # url - the url of the contract page
-  # Returns a hash containing the contract information.
-  # See test/lib/scrapers/<%= @agency.downcase %>/<%= @agency.downcase %>_crawler_test.rb for data format.
-  def contract_hash(url)
-    raise "Not implemented"
-    attrs = Wombat.crawl do
-      base_url url
-      path ""
-      vendor_name "css="
-      reference_number "css="
-      effective_date "css="
-      end
-      description do
-        main "css="
-      end
-      raw_contract_period "css="
-      value "css=" do |amount|
-        Monetize.parse(amount).cents / 100
-      end
-      comments do
-        main "css="
-      end
-    end
-    attrs["url"] = url
-    attrs
-  end
-
-  # Return an Array with the urls the parser needs to visit to scrape all
-  # contracts in this quarter.
-  def contract_urls(quarter)
-    raise "Not implemented"
-    urls = Wombat.crawl do
-      # The base_url and path must form the url to the given contracts page for this quarter
-      base_url ""
-      path     ""
-      # Link selector goes here.
-    end
-  end
-end
