+++ lib/generators/scraper/templates/scraper.rb.erb
@@ -1,0 +1,11 @@
+class Scrapers::<%= @agency.camelize %>::Scraper < Scrapers::ContractScraper
+
+  def contracts(range = 0..-1)
+
+  end
+
+  def self.available_quarters
+
+  end
+end
+
