+++ lib/scrapers/dfo/scraper.rb
@@ -1,7 +1,13 @@
-class Scrapers::Dfo::DfoCrawler < Scrapers::ContractCrawler
+class Scrapers::Dfo::Scraper < Scrapers::ContractScraper
 
-  protected
-  
+  def contracts(range = 0..-1)
+    contract_urls(quarter)[range].collect do |url|
+      flatten_to_json(contract_hash(url))
+    end
+  end
+
+  private
+
   def contract_hash(url)
     attrs = Wombat.crawl do
       base_url url
